+++ Portal/PortalStackServiceContainerFactory.php
@@ -4,11 +4,11 @@ declare(strict_types=1);
 
 namespace Heptacom\HeptaConnect\Core\Portal;
 
+use Heptacom\HeptaConnect\Core\Portal\Contract\PortalNodeContainerFacadeContract;
 use Heptacom\HeptaConnect\Core\Portal\Contract\PortalRegistryInterface;
 use Heptacom\HeptaConnect\Core\Portal\Contract\PortalStackServiceContainerBuilderInterface;
 use Heptacom\HeptaConnect\Portal\Base\StorageKey\Contract\PortalNodeKeyInterface;
 use Heptacom\HeptaConnect\Storage\Base\Contract\StorageKeyGeneratorContract;
-use Psr\Container\ContainerInterface;
 
 class PortalStackServiceContainerFactory
 {
@@ -18,6 +18,9 @@ class PortalStackServiceContainerFactory
 
     private StorageKeyGeneratorContract $storageKeyGenerator;
 
+    /**
+     * @var PortalNodeContainerFacadeContract[]
+     */
     private array $portalContainers = [];
 
     public function __construct(
@@ -30,22 +33,23 @@ class PortalStackServiceContainerFactory
         $this->storageKeyGenerator = $storageKeyGenerator;
     }
 
-    public function create(PortalNodeKeyInterface $portalNodeKey): ContainerInterface
+    public function create(PortalNodeKeyInterface $portalNodeKey): PortalNodeContainerFacadeContract
     {
         $key = $this->storageKeyGenerator->serialize($portalNodeKey);
         $result = $this->portalContainers[$key] ?? null;
 
-        if ($result instanceof ContainerInterface) {
+        if ($result instanceof PortalNodeContainerFacadeContract) {
             return $result;
         }
 
-        $result = $this->portalStackServiceContainerBuilder->build(
+        $container = $this->portalStackServiceContainerBuilder->build(
             $this->portalRegistry->getPortal($portalNodeKey),
             $this->portalRegistry->getPortalExtensions($portalNodeKey),
             $portalNodeKey
         );
-        $result->compile();
-        $this->portalContainers[$key] = $result;
+        $container->compile();
+        $result = new PortalNodeContainerFacade($container);
+        $this->portalContainers[$key] = result;
 
         return $result;
     }